Redoing changes that could not be recovered due to the events of The Hammer Part 3
An unexpected late-night push on the eve of presentations resulted in our entire web page breaking. Alas, this disastrous consequence of such a seemingly innocent action was discovered by neither the pusher, nor the rest of us (who were undoubtedly asleep at the time). It was not until we arrived for presentations in the morning that measures could be taken in an attempt to recover a working web page. Despite the brave efforts of the group (and Kirby), git revert commands were unable to yield a solution. Thus, it was with great regret that I initiated The Hammer Part 3: The Blacksmith's Anvil (a git push --force from my machine, which still had the old version), a blow so powerful and terrible (Kirby let us know that we should never do this: but to do it in this situation to save our project) that it tore a hole in the very fabric of time and space (the history of our repo was changed). While the day was saved, our heroes could do naught but stand and witness the devastation such an act had caused (we lost a few commits made after the version we force pushed). While the world would go on to rebuild and move on from such a catastrophe (me making this push), the world would never forget that devilish day, The Push to End All Pushes, The Sundering, The Fifth Turmoil, The Night of the Dark Commit.